backend/requirements.txt
asgiref==3.2.10
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.0.4
Pillow==8.3.1
PyJWT==2.1.0
psycopg2-binary==2.8.6
pytz==2023.3
sqlparse==0.4.3
